Fixed GetCurrentMap

+++ src/field_effect.c
@@ -3928,7 +3928,7 @@ static void Task_MoveDeoxysRock(u8 taskId)
 
 u32 GetCurrentMap(void)//Used for PSF Flying Blind
 {
-    u8 map = (gSaveBlock1Ptr->location.mapGroup << 8) | gSaveBlock1Ptr->location.mapNum;
+    u16 map = (gSaveBlock1Ptr->location.mapGroup << 8) | gSaveBlock1Ptr->location.mapNum;
 
     gSpecialVar_Result = map;
     return map;
